内容:按最新的工程，application、domain、repository三个可以生成代码可用

+++ microservice-autocode/src/main/java/com/liujun/micro/autocode/generator/builder/operator/code/GenerateJavaInterfaceConstant.java
@@ -234,7 +234,7 @@ public class GenerateJavaInterfaceConstant {
    * @param tableColumn 表列信息
    * @return 类型的最大长度
    */
-  private Integer maxLength(DatabaseTypeEnum dbType, TableColumnDTO tableColumn) {
+  private Long maxLength(DatabaseTypeEnum dbType, TableColumnDTO tableColumn) {
     if (tableColumn.getDataLength() != null && tableColumn.getDataLength() != 0) {
       return tableColumn.getDataLength();
     } else {
